+++ resources/assets/js/components/enso/vueforms/VueFormSs.vue
@@ -1,6 +1,7 @@
 <template>
 
     <vue-form :data="data"
+        :locale="$store.state.user.preferences.global.lang"
         v-if="data"
         ref="form">
         <template v-for="field in customFields"
@@ -28,6 +29,16 @@ export default {
         params: {
             type: Array,
             required: true,
+        },
+        locale: {
+            type: String,
+            default() {
+                let locale = 'en';
+                try {
+                    locale = this.$store.state.user.preferences.global.lang;
+                } catch (e) {}
+                return locale;
+            },
         },
     },